Add hex and underscore support to integer literals

+++ src/FLang.Frontend/Ast/Expressions/IntegerLiteralNode.cs
@@ -1,15 +1,16 @@
+using System.Numerics;
 using FLang.Core;
 
 namespace FLang.Frontend.Ast.Expressions;
 
 public class IntegerLiteralNode : ExpressionNode
 {
-    public IntegerLiteralNode(SourceSpan span, long value, string? suffix = null) : base(span)
+    public IntegerLiteralNode(SourceSpan span, BigInteger value, string? suffix = null) : base(span)
     {
         Value = value;
         Suffix = suffix;
     }
 
-    public long Value { get; }
+    public BigInteger Value { get; }
     public string? Suffix { get; }
 }
